refactor: ♻️ 函数解耦
