+++ tsconfig.json
@@ -30,6 +30,6 @@
     "scripts/build-cli.ts",
     "bin/*.ts",
     "lib/*.ts",
-    "lib/src/*.ts"
-  ]
+    "lib/utils/*.ts"
+, "lib/genFile/relations.ts", "lib/genFile/outputFile.ts"  ]
 }
